Refactor DocBlockResolverAwareTrait extract logic into class
